Refactor layout and components for "Rick Roll Call" app; implement audio generation and Twilio call features. Add About section and AudioPlayer component. Update metadata and remove unused files.
